Move reading to Record
